% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{human_equilibrium_vivax_full_het}
\alias{human_equilibrium_vivax_full_het}
\title{Vivax equilibrium solution with biting heterogeneity}
\usage{
human_equilibrium_vivax_full_het(EIR, ft, p, age, h = NULL)
}
\arguments{
\item{EIR}{EIR for adults, in units of infectious bites per person per year}

\item{ft}{proportion of clinical cases effectively treated}

\item{p}{vector of model parameters}

\item{age}{vector of age groups, in units of years}

\item{h}{a list of Gauss-Hermite nodes and associated weights, used for
integrating over heterogeneity in biting. See \code{?gq_normal} for an
example.}
}
\description{
Returns the vivax equilibrium states for the model of White et al.,
  (2018).

  This function does account for biting heterogeneity and return the 
  equilibria listed over heterogeneity - to summarise and calculate FOIM 
  please use \code{human_equilibrium_vivax_summarise_het()}.
  
  This version uses Michael White's code as a template. I've removed the omega 
  standardisation for consistency with other equiilbria.
}
