% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{human_equilibrium_vivax_summarise}
\alias{human_equilibrium_vivax_summarise}
\title{Vivax equilibrium solution}
\usage{
human_equilibrium_vivax_summarise(eq, p)
}
\arguments{
\item{eq}{the equilibrium solution calcualted by human_equilibrium_vivax_full_het}

\item{p}{a parameter set generated by get_parameters() and complementary functions
in malariasimulation}
}
\description{
Returns the vivax equilibrium states for the model of White et al.,
  (2018).
}
