% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{malariaEqVivax_file}
\alias{malariaEqVivax_file}
\title{Load system file for this package}
\usage{
malariaEqVivax_file(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Load and return file from within the inst folder of this
  package.
}
